Update requirements.yaml to cover new functions
These tools are all used within our environment so need to be available in this
repository.
I'm pinning DebOps to the stable branch in git to try and reduce churn a
little. It is still a git repo (so not based on clean releases) but I hope not
quite as wild to live in as the master branch.