Move to latest version of OpenStack SDK
Moving to new version has resulted in some compatibility adjustments in
playbooks but was long overdue. It allows me to use the latest openstack cli
client so has been worth the effort.

There is one problem with the change - the return from instances being created
appears to leave no easy way to access the public IP, with the api now
returning a dict of addresses and making me sort through them.